requirements.txt
torch>=2.0
torchdata                                         # 新增
pytorch-lightning>=2.2
hydra-core
nltk
w3lib
tqdm
scikit-learn
transformers>=4.38
sentencepiece
